# Debloat: invert_limb and mpoly_heap methods