Size of libflint.so:

    Old: 46470280
    New: 41542920  (10% smaller)

Build time (-j8):

    Old:
    real	1m7,376s
    user	6m32,535s
    sys	0m51,132s

    New:
    real	1m5,485s      (3% faster)
    user	6m20,262s
    sys	0m50,772s