@fredrik-johansson 2022-05-19
Fixes issue #1123.