@albinahlback 2022-05-19
Perhaps the same should go for Conway polynomials? Now it's in a header, and should thus create the same problem (although not at this level).